requirements.txt
python-dotenv
openai
langchain
langchain_core
langgraph
langchain_openai
langchain_community
faiss-cpu
llama_index